Sonja 6 19
This code is still a mess--it works in IDLE for me, but I can't get past 4 points on zybooks. Just to look at for ideas if it helps you at all. I have been looking at your code too--on my end I get 14/20 in zybooks using yours. I'm totally stumped as to the couple of errors your getting, and your code looks great, so I'm at a total loss.